refactor: extract runtime-safe schema observation helpers
Split runtime-safe schema observation/configuration helpers out of the tooling-side
sampling module and rewire runtime plus shared callers onto the new boundary.
