+++ polylogue/schemas/registry.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 
+from polylogue.schemas.observation import fingerprint_hash as _fingerprint_hash
 from polylogue.schemas.runtime_registry import SCHEMA_DIR, canonical_schema_provider
 from polylogue.schemas.runtime_registry import SchemaRegistry as RuntimeSchemaRegistry
-from polylogue.schemas.sampling import fingerprint_hash as _fingerprint_hash
 from polylogue.schemas.tooling_registry import (
     ClusterManifest,
     PropertyChange,
